Fix outstanding import cycles
